+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/HandlerKeyTest.java
@@ -1,9 +1,8 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.interface21.web.bind.annotation.RequestMethod;
-import com.interface21.webmvc.servlet.mvc.HandlerKey;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
